Expressões com tipo Double
Melhorias para aceitar números do tipo Double para realizações de expressões com resultados mais específicos.

+++ src/agents/Multiplicacao.java
@@ -1,14 +1,10 @@
 package agents;
 
-import behaviours.BehaviourSoma;
-import behaviours.BehaviourControle;
 import behaviours.BehaviourMultiplicacao;
-import behaviours.BehaviourSubtracao;
 import jade.core.Agent;
-import jade.core.behaviours.Behaviour;
 
 public class Multiplicacao extends Agent{
-    private int valores[];
+    private double valores[];
 	
 	private static final long serialVersionUID = 2L;
