+++ src/agents/Subtracao.java
@@ -1,13 +1,10 @@
 package agents;
 
-import behaviours.BehaviourSoma;
-import behaviours.BehaviourControle;
 import behaviours.BehaviourSubtracao;
 import jade.core.Agent;
-import jade.core.behaviours.Behaviour;
 
 public class Subtracao extends Agent{
-    private int valores[];
+    private double valores[];
 	
 	private static final long serialVersionUID = 2L;
